tests/wpt/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox_005.htm

<!DOCTYPE html>
<html>
<head>
    <title>HTML5 Sandbox: Block script execution inside iframe with sandbox attribute.</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type" />
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://dev.w3.org/html5/spec/Overview.html#sandboxed-scripts-browsing-context-flag" />
    <meta name="assert" content="Block script execution inside iframe with sandbox attribute." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script type="text/javascript">

        var t = async_test("Block script execution inside iframe with sandbox attribute.");

        function callback(event)
        {
            t.step(function(){
                assert_true('sandbox' in document.createElement('iframe'));
                assert_true(!event);
            });
            t.done();
        }

        var timer = setTimeout(callback, 4000);
        window.addEventListener("message", callback, false);
    </script>
    <iframe src="support/iframe_sandbox_001.htm" sandbox style="display: none"></iframe>
    <div id=log></div>
</body>
</html>
