LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox_005-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://localhost:8800/html/semantics/embedded-content/the-iframe-element/support/iframe_sandbox_001.htm' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.

PASS Block script execution inside iframe with sandbox attribute.
